test examples created
