+++ examples/__init__.py
@@ -1,0 +1,1 @@
+# Examples package for PyGIP experiments and demonstrations
